Preparacion para release

+++ Interface/ModifyUserUI.cs
@@ -1,15 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel;
-using System.Drawing;
-using System.Data;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 using Domain;
-using Persistence;
-using Logic;
+using Exceptions;
 using System.Globalization;
 
 namespace Interface
@@ -18,31 +11,40 @@ namespace Interface
     {
         private Instance instance;
         private List<User> users { get; set; }
-
         public ModifyUserUI()
         {
             instance = Instance.GetInstance;
             InitializeComponent();
             LoadUserList();
             LoadUsersToDataGrid();
-        }     
+        }
 
         private void btnModify_Click(object sender, EventArgs e)
         {
-            var selectedRow = this.dataGridViewUsers.CurrentCell.RowIndex;
-            var selectedUser = this.dataGridViewUsers.Rows[selectedRow].Tag;
-            User user = (User)selectedUser;
+            if (txtActualPassword.Text.Length == 0)
+            {
+                MessageBox.Show("Debe seleccionar un usuario para cambiar su contraseña.", "Informacion", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
+            else
+            {
+                var selectedRow = this.dataGridViewUsers.CurrentCell.RowIndex;
+                var selectedUser = this.dataGridViewUsers.Rows[selectedRow].Tag;
+                User user = (User)selectedUser;
                 try
                 {
                     instance.UserHandler.ModifyPassword(user.mail, txtNewPassword.Text);
                     MessageBox.Show("Modificación exitosa.", "Informacion", MessageBoxButtons.OK, MessageBoxIcon.Information);
                     txtActualPassword.Text = "";
                     txtNewPassword.Text = "";
+                    LoadUserList();
+                    LoadUsersToDataGrid();
                 }
-                catch (Exception)
+                catch (UserException ex)
                 {
-                    MessageBox.Show("Contraseña nueva invalida, ingrese una correcta.", "Informacion", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    String msgError = ex.Message;
+                    MessageBox.Show(msgError, "Informacion", MessageBoxButtons.OK, MessageBoxIcon.Information);
                 }
+            }
         }
 
         private void BtnSelect_Click(object sender, EventArgs e)
@@ -63,19 +65,14 @@ namespace Interface
         private void LoadUserList()
         {
             users = new List<User>();
-            using (ContextDB context = new ContextDB())
+            List<User> allUsers = instance.UserHandler.LoadUsers();
+            foreach (User user in allUsers)
             {
-                foreach (User user in context.Users.ToList())
+                if (user.active)
                 {
-                    if (user.active)
-                    {
-                        users.Add(user);
-                    }
-                        
+                    users.Add(user);
                 }
-
             }
-
         }
         private void LoadUsersToDataGrid()
         {
@@ -83,10 +80,8 @@ namespace Interface
             this.dataGridViewUsers.Rows.Clear();
             foreach (User user in users)
             {
-                
                 var rowIndex = this.dataGridViewUsers.Rows.Add(user.mail, user.name, user.surname);
                 this.dataGridViewUsers.Rows[rowIndex].Tag = user;
-
             }
         }
 
@@ -95,9 +90,20 @@ namespace Interface
             var selectedRow = this.dataGridViewUsers.CurrentCell.RowIndex;
             var selectedUser = this.dataGridViewUsers.Rows[selectedRow].Tag;
             User user = (User)selectedUser;
-            instance.UserHandler.DeleteUser(user.mail);
-            LoadUserList();
-            LoadUsersToDataGrid();
+            if (user.mail.Equals(instance.Session.user.mail))
+            {
+                MessageBox.Show("No puede eliminarse a usted mismo.", "Información", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
+            else
+            {
+                DialogResult result = MessageBox.Show("Desea eliminar a " + user.mail + "?", "Confirmacion", MessageBoxButtons.YesNo, MessageBoxIcon.Exclamation);
+                if (result == System.Windows.Forms.DialogResult.Yes)
+                {
+                    instance.UserHandler.DeleteUser(user.mail);
+                    LoadUserList();
+                    LoadUsersToDataGrid();
+                }
+            }
         }
     }
 }
